Renamed org.scalatest.check package as org.scalatestplus.scalacheck.
