+++ scalatestPlusScalaCheck/src/test/scala/org/scalatest/check/PropertyCheckConfigurationHelperSuite.scala
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
-package check
+package org.scalatestplus
+package scalacheck
 
 import org.scalactic.anyvals._
 
-class PropertyCheckConfigurationHelperSuite extends FunSuite with Matchers {
+class PropertyCheckConfigurationHelperSuite extends org.scalatest.funsuite.AnyFunSuite with org.scalatest.Matchers {
 
-  import org.scalatest.check.ScalaCheckConfiguration._
+  import org.scalatestplus.scalacheck.ScalaCheckConfiguration._
 
   val DefaultMinSuccessful: PosInt = 9
   val PassedMinSuccessful: PosInt = 3
